feat: mini posts
Add Mini posts data to sidebar. A GraphQL query is made to fetch the
latest 5 posts and is passed down to the Sidebar which contain the
MiniPosts component.

The query is made at the TemplateWrapper as GraphQL queries in Gatsby,
can only be made at either Layouts or Templates, this is then passed
down as props to the Sidebar all the way to MiniPosts to be rendered.