fix(navbar): remove projects tab
Remove Projects tab as it led to a crashing Projects Page. Github API integration still fails in
production due to Github revoking API token when exposed and published to Github. This will need an
alternative to load Github Projects on the blog, or remove it all together and include it as a
separate site